remove webpack, organize folders
