+++ modules/arrayAnalyze.js
