+++ package.json
@@ -4,10 +4,6 @@
   "description": "Test suit build with JEST.",
   "main": "index.js",
   "scripts": {
-    "build": "webpack --mode development",
-    "build:watch": "webpack -w --mode development",
-    "build:prod": "webpack --mode production",
-    "go": "webpack-dev-server --host 0.0.0.0 --port 7000",
     "stylelint": "npx stylelint ./src/css/*.css",
     "eslint": "npx eslint ./src/*.js",
     "test": "jest"
@@ -26,23 +22,10 @@
     "@babel/core": "^7.11.4",
     "@babel/preset-env": "^7.11.0",
     "babel-loader": "^8.1.0",
-    "css-loader": "^4.2.2",
-    "css-minimizer-webpack-plugin": "^1.1.2",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb-base": "^14.1.0",
-    "eslint-plugin-import": "^2.20.2",
-    "file-loader": "^6.0.0",
-    "html-loader": "^1.3.0",
-    "html-webpack-plugin": "^4.3.0",
     "jest": "^26.4.2",
-    "mini-css-extract-plugin": "^0.11.0",
-    "sass": "^1.26.10",
-    "sass-loader": "^10.0.1",
     "stylelint": "^13.3.3",
     "stylelint-config-standard": "^20.0.0",
     "stylelint-csstree-validator": "^1.8.0",
-    "stylelint-scss": "^3.17.2",
-    "webpack": "^4.44.1",
-    "webpack-dev-server": "^3.11.0"
+    "stylelint-scss": "^3.17.2"
   }
 }
