+++ test/index.test.js
@@ -1,4 +1,4 @@
-import capitalize from './index';
+import capitalize from '../modules/index';
 
 it('Should capitalize the first letter', () => {
   expect(capitalize('stuff')).toBe('Stuff');
